+++ src/index.d.ts
@@ -4,6 +4,7 @@ export { useCountUp } from "./composables/useCountUp";
 export { useTypewriter } from "./composables/useTypewriter";
 export { useSSE } from "./composables/useSSE";
 export { useClipboard } from "./composables/useClipboard";
+export { useRandomWords } from "./composables/useRandomWords";
 export { useEventBus } from "./wrapper/useEventBus";
 export { useFetch } from "./wrapper/useFetch";
 export { useAuthGuard } from "./wrapper/useAuthGuard";